Separate Voter Info into its own Package
